feat: organization logotypes
