+++ backend/src/controllers/messaging-settings.controller.ts
@@ -9,7 +9,7 @@ import { getApiBase, MUNICIPALITY_ID } from '@config';
 import { MessagingSettings } from '@/data-contracts/messaging-settings/data-contracts';
 import { RequestWithUser } from '@interfaces/auth.interface';
 import { MessagingSettingsApiResponse } from '@/responses/messaging-settings.response';
-import { Logotype } from '@interfaces/logotypes.interface';
+import { structureLogotypeData } from '@services/messaging-settings.service';
 
 @Controller()
 @UseBefore(authMiddleware)
@@ -38,28 +38,7 @@ export class MessagingSettingsController {
       const filter = `values.key = 'host' and values.value = '${hostName}'`;
       const url = `${this.baseUrl}?filter=${encodeURIComponent(filter)}`;
       const result = await this.apiService.get<MessagingSettings[]>({ url }, req.user);
-
-      const LOGOTYPE_KEYS = ['host', 'display_name', 'logotype_lightmode', 'logotype_darkmode'] as const;
-      type LogotypeKey = (typeof LOGOTYPE_KEYS)[number];
-
-      const hostLogotype: Logotype[] = result.data.map(d => {
-        const fields = d.values.reduce<Partial<Record<LogotypeKey, string>>>((acc, v) => {
-          if ((LOGOTYPE_KEYS as readonly string[]).includes(v.key)) {
-            acc[v.key as LogotypeKey] = v.value;
-          }
-          return acc;
-        }, {});
-
-        return {
-          id: d.id,
-          host: fields.host ?? '',
-          display_name: fields.display_name ?? '',
-          logotype_lightmode: fields.logotype_lightmode,
-          logotype_darkmode: fields.logotype_darkmode,
-        };
-      });
-
-      return res.send({ message: 'success', data: hostLogotype ?? [] });
+      return res.send({ message: 'success', data: structureLogotypeData(result.data) ?? [] });
     } catch (error) {
       logger.error('Error getting host logotypes', error);
       throw new HttpException(500, 'Could not get host logotypes');
